fix(e2e): break circular import by extracting shutdown coordination to shutdown.py
runner.py -> tier_action_builder -> subtest_executor.__getattr__ -> parallel_executor
-> rate_limit -> (was) runner caused ImportError on 'InfrastructureFailureError' when
runner was only partially initialized.

Extract ShutdownInterruptedError, is_shutdown_requested, and request_shutdown into a
new lightweight scylla.e2e.shutdown module. All lazy importers (rate_limit, stages,
state_machine, subtest_state_machine, tier_state_machine, experiment_state_machine,
parallel_executor, parallel_tier_runner) now import from shutdown instead of runner.
runner.py re-exports all three symbols for backward compatibility.
