+++ src/scylla/e2e/runner.py
@@ -56,37 +56,20 @@ logger = logging.getLogger(__name__)
 # Checkpoint status constant (kept as string for JSON serialization compatibility)
 _STATUS_RUNNING = "running"
 
-# Global shutdown coordination
-_shutdown_requested = False
+# Shutdown coordination — re-exported from shutdown.py for backward compatibility.
+# Callers that only need these symbols should import from scylla.e2e.shutdown directly
+# to avoid the circular import: runner -> ... -> rate_limit -> runner.
+from scylla.e2e.shutdown import (  # noqa: E402
+    ShutdownInterruptedError,
+    is_shutdown_requested,
+    request_shutdown,
+)
 
-
-class ShutdownInterruptedError(Exception):
-    """Raised when an in-progress stage is interrupted by a shutdown signal (Ctrl+C).
-
-    Unlike a generic Exception, this is caught separately by StateMachine.advance_to_completion()
-    so the run is NOT marked as FAILED.  The run state stays at its last successfully
-    checkpointed value, allowing clean resume on the next invocation.
-    """
-
-
-def request_shutdown() -> None:
-    """Request graceful shutdown of the experiment.
-
-    This is typically called by signal handlers (SIGINT, SIGTERM).
-    """
-    global _shutdown_requested
-    _shutdown_requested = True
-    logger.warning("Graceful shutdown requested")
-
-
-def is_shutdown_requested() -> bool:
-    """Check if shutdown has been requested.
-
-    Returns:
-        True if shutdown is requested, False otherwise
-
-    """
-    return _shutdown_requested
+__all__ = [
+    "ShutdownInterruptedError",
+    "is_shutdown_requested",
+    "request_shutdown",
+]
 
 
 @dataclass
